neuronotes_be/requirements.txt
annotated-types==0.7.0
pydantic==2.10.6
pydantic_core==2.27.2
typing_extensions==4.12.2
fastapi==0.104.1
uvicorn==0.24.0
moviepy==1.0.3
requests==2.31.0
python-multipart==0.0.6
gdown==5.1.0
google-auth-oauthlib==1.2.0
google-auth==2.28.1
google-api-python-client==2.120.0
google-generativeai==0.3.2
jinja2==3.1.3
aiofiles==23.2.1
httpx==0.27.0
python-dotenv==1.0.0
youtube-transcript-api==0.6.2